Commit from JBoss Tools

+++ src/main/java/com/hout/business/dao/SuggestionDao.java
@@ -2,7 +2,7 @@ package com.hout.business.dao;
 
 import com.hout.domain.entities.Suggestion;
 
-public interface SuggestionDao extends GenericDao<Suggestion, Integer>{
+public interface SuggestionDao extends GenericDao<Suggestion, Long>{
 
 	public Suggestion findById(long id);
